restructure files to be more modular
