+++ components/disclaimer.module.css
@@ -1,0 +1,4 @@
+.disclaimer {
+  font-size: 0.8em;
+  font-style: italic;
+}
