+++ components/disclaimer.tsx
@@ -1,0 +1,11 @@
+import styles from "./disclaimer.module.css";
+
+export default function Disclaimer() {
+  return (
+    <>
+      <p className={styles.disclaimer}>
+        Images are used on this site for educational purposes only.
+      </p>
+    </>
+  );
+}
