+++ components/footer.module.css
@@ -1,0 +1,14 @@
+.footer {
+  width: 100%;
+  height: 100px;
+  border-top: 1px solid #b4747d;
+}
+
+.footer a {
+  color: #b4747d;
+  font-weight: bold;
+}
+
+.iconLink {
+  text-decoration: none;
+}
