+++ utils/statuses.ts
@@ -1,0 +1,21 @@
+const statuses = [
+  { code: 100, title: "Continue" },
+  { code: 101, title: "Switching Protocols" },
+  { code: 200, title: "OK" },
+  { code: 202, title: "Accepted" },
+  { code: 300, title: "Multiple Choices" },
+  { code: 301, title: "Moved Permanently" },
+  { code: 400, title: "Bad Request" },
+  { code: 402, title: "Payment Required" },
+  { code: 403, title: "Forbidden" },
+  { code: 406, title: "Not Acceptable" },
+  { code: 407, title: "Proxy Authentication Required" },
+  { code: 409, title: "Conflict" },
+  { code: 410, title: "Gone" },
+  { code: 418, title: "I'm a teapot" },
+  { code: 426, title: "Upgrade required" },
+  { code: 429, title: "Too many requests" },
+  { code: 503, title: "Service unavailable" },
+];
+
+export default statuses;
